utilities/video-processor/requirements.txt
moviepy==1.0.3
decorator==4.4.2
tqdm==4.62.3
imageio==2.34.0
imageio-ffmpeg==0.4.5
google-cloud-storage
google-cloud-firestore
Pillow==9.5.0
setuptools>=65.5.0
wheel
Flask==3.0.0
gunicorn==21.2.0
google-auth>=2.17.3